+++ onecool_os/assets/master/__init__.py
@@ -1,0 +1,17 @@
+"""Asset Master user-owned metadata layer."""
+
+from onecool_os.assets.master.loader import AssetMasterLoader
+from onecool_os.assets.master.merge import PROTECTED_IDENTITY_FIELDS
+from onecool_os.assets.master.merge import merge_asset_master
+from onecool_os.assets.master.models import AssetMasterLoadResult
+from onecool_os.assets.master.models import AssetMasterRecord
+from onecool_os.assets.master.validation import AssetMasterError
+
+__all__ = [
+    "AssetMasterError",
+    "AssetMasterLoadResult",
+    "AssetMasterLoader",
+    "AssetMasterRecord",
+    "PROTECTED_IDENTITY_FIELDS",
+    "merge_asset_master",
+]